stage-2_3DGS/depthsplat/requirements.txt
beartype
colorama
colorspacious
dacite
e3nn
einops
hydra-core
jaxtyping
lpips
matplotlib
moviepy
numpy
opencv_python
Pillow
plyfile
pytorch_lightning
scikit-image
sk-video
tabulate
tqdm
wandb
xformers